[AIRFLOW-5741] Move Cloud Natural Language to providers
