+++ airflow/providers/google/cloud/hooks/natural_language.py
