+++ tests/providers/google/cloud/hooks/test_natural_language.py
@@ -22,7 +22,7 @@ from typing import Any, Dict
 
 from google.cloud.language_v1.proto.language_service_pb2 import Document
 
-from airflow.gcp.hooks.natural_language import CloudNaturalLanguageHook
+from airflow.providers.google.cloud.hooks.natural_language import CloudNaturalLanguageHook
 from tests.compat import mock
 from tests.gcp.utils.base_gcp_mock import mock_base_gcp_hook_no_default_project_id
 
@@ -42,11 +42,12 @@ class TestCloudNaturalLanguageHook(unittest.TestCase):
             self.hook = CloudNaturalLanguageHook(gcp_conn_id="test")
 
     @mock.patch(
-        "airflow.gcp.hooks.natural_language.CloudNaturalLanguageHook.client_info",
+        "airflow.providers.google.cloud.hooks.natural_language.CloudNaturalLanguageHook.client_info",
         new_callable=mock.PropertyMock
     )
-    @mock.patch("airflow.gcp.hooks.natural_language.CloudNaturalLanguageHook._get_credentials")
-    @mock.patch("airflow.gcp.hooks.natural_language.LanguageServiceClient")
+    @mock.patch("airflow.providers.google.cloud.hooks.natural_language.CloudNaturalLanguageHook."
+                "_get_credentials")
+    @mock.patch("airflow.providers.google.cloud.hooks.natural_language.LanguageServiceClient")
     def test_language_service_client_creation(self, mock_client, mock_get_creds, mock_client_info):
         result = self.hook.get_conn()
         mock_client.assert_called_once_with(
@@ -57,7 +58,7 @@ class TestCloudNaturalLanguageHook(unittest.TestCase):
         self.assertEqual(self.hook._conn, result)
 
     @mock.patch(  # type: ignore
-        "airflow.gcp.hooks.natural_language.CloudNaturalLanguageHook.get_conn",
+        "airflow.providers.google.cloud.hooks.natural_language.CloudNaturalLanguageHook.get_conn",
         **{"return_value.analyze_entities.return_value": API_RESPONSE}  # type: ignore
     )
     def test_analyze_entities(self, get_conn):
@@ -70,7 +71,7 @@ class TestCloudNaturalLanguageHook(unittest.TestCase):
         )
 
     @mock.patch(  # type: ignore
-        "airflow.gcp.hooks.natural_language.CloudNaturalLanguageHook.get_conn",
+        "airflow.providers.google.cloud.hooks.natural_language.CloudNaturalLanguageHook.get_conn",
         **{"return_value.analyze_entity_sentiment.return_value": API_RESPONSE}
     )
     def test_analyze_entity_sentiment(self, get_conn):
@@ -83,7 +84,7 @@ class TestCloudNaturalLanguageHook(unittest.TestCase):
         )
 
     @mock.patch(  # type: ignore
-        "airflow.gcp.hooks.natural_language.CloudNaturalLanguageHook.get_conn",
+        "airflow.providers.google.cloud.hooks.natural_language.CloudNaturalLanguageHook.get_conn",
         **{"return_value.analyze_sentiment.return_value": API_RESPONSE}
     )
     def test_analyze_sentiment(self, get_conn):
@@ -96,7 +97,7 @@ class TestCloudNaturalLanguageHook(unittest.TestCase):
         )
 
     @mock.patch(  # type: ignore
-        "airflow.gcp.hooks.natural_language.CloudNaturalLanguageHook.get_conn",
+        "airflow.providers.google.cloud.hooks.natural_language.CloudNaturalLanguageHook.get_conn",
         **{"return_value.analyze_syntax.return_value": API_RESPONSE}
     )
     def test_analyze_syntax(self, get_conn):
@@ -109,7 +110,7 @@ class TestCloudNaturalLanguageHook(unittest.TestCase):
         )
 
     @mock.patch(  # type: ignore
-        "airflow.gcp.hooks.natural_language.CloudNaturalLanguageHook.get_conn",
+        "airflow.providers.google.cloud.hooks.natural_language.CloudNaturalLanguageHook.get_conn",
         **{"return_value.annotate_text.return_value": API_RESPONSE}
     )
     def test_annotate_text(self, get_conn):
@@ -127,7 +128,7 @@ class TestCloudNaturalLanguageHook(unittest.TestCase):
         )
 
     @mock.patch(  # type: ignore
-        "airflow.gcp.hooks.natural_language.CloudNaturalLanguageHook.get_conn",
+        "airflow.providers.google.cloud.hooks.natural_language.CloudNaturalLanguageHook.get_conn",
         **{"return_value.classify_text.return_value": API_RESPONSE}
     )
     def test_classify_text(self, get_conn):
