+++ tests/providers/google/cloud/operators/test_natural_language_system.py
